--- a/src/handlers/javadoc.rs
+++ b/src/handlers/javadoc.rs
@@ -6,7 +6,7 @@ use regex::{Regex, RegexBuilder};
 use std::io;
 use std::io::{BufRead, BufWriter, Write};
 use std::path::Path;
-use std::rc::Rc;
+use std::sync::Arc;
 
 use crate::handlers::InputOutputHelper;
 use crate::config;
@@ -14,11 +14,11 @@ use crate::config;
 const HEADER_LINES_TO_CHECK: i32 = 15;
 
 pub struct Javadoc {
-    config: Rc<config::Config>,
+    config: Arc<config::Config>,
 }
 
 impl Javadoc {
-    pub fn boxed(config: &Rc<config::Config>) -> Box<dyn super::Processor> {
+    pub fn boxed(config: &Arc<config::Config>) -> Box<dyn super::Processor + Send + Sync> {
         Box::new(Self { config: config.clone() })
     }
 
@@ -150,7 +150,7 @@ mod tests {
 
     #[test]
     fn test_filter_html() {
-        let cfg = Rc::new(config::Config::empty(1704106800, false));
+        let cfg = Arc::new(config::Config::empty(1704106800, false));
         let h = Javadoc::boxed(&cfg);
 
         assert!( h.filter(Path::new("/some/path/page.html")).unwrap());
@@ -164,7 +164,7 @@ mod tests {
 
     #[test]
     fn test_process_line() {
-        let config = Rc::new(config::Config::empty(1704106800, false));
+        let config = Arc::new(config::Config::empty(1704106800, false));
         let h = Javadoc { config };
         let plu = |s| h.process_line(s).unwrap();