--- a/tests/test_handlers/mod.rs
+++ b/tests/test_handlers/mod.rs
@@ -11,7 +11,7 @@ use anyhow::Result;
 use std::fs;
 use std::io::{Read, Write};
 use std::path::Path;
-use std::rc::Rc;
+use std::sync::Arc;
 
 use add_determinism::config;
 use add_determinism::handlers;
@@ -25,7 +25,7 @@ fn make_handler(
 
     init();
 
-    let cfg = Rc::new(config::Config::empty(source_date_epoch, check));
+    let cfg = Arc::new(config::Config::empty(source_date_epoch, check));
     let mut handler = func(&cfg);
     handler.initialize()?;
     Ok(handler)
@@ -34,7 +34,7 @@ fn make_handler(
 struct Trivial {}
 
 impl Trivial {
-    pub fn boxed() -> Box<dyn handlers::Processor> {
+    pub fn boxed() -> Box<dyn handlers::Processor + Send + Sync> {
         Box::new(Self {})
     }
 }
@@ -130,7 +130,7 @@ fn test_inode_map() {
 fn test_inode_map_2() {
     let (dir, _input) = prepare_dir("tests/cases/testrelro.a").unwrap();
 
-    let cfg = Rc::new(config::Config::empty(111, false));
+    let cfg = Arc::new(config::Config::empty(111, false));
     let ar = handlers::ar::Ar::boxed(&cfg);
 
     let handlers = vec![ar];