Originally, add-det used processes, because the pyc worker was using an embedded python interpreter. At the time multiple embedded parallel Python VMs were considered experimental so threads were not an option. But since pyc was rewritten in Rust, there is no good reason not to use threads. Doing this allows the low-level code to serialize/deserialize the jobs and results and manipulate sockets to be dropped. All unsafe code is now gone.

Timings are similar:
```console
$ time SOURCE_DATE_EPOCH=1748901600 target/release/add-determinism.old --check /usr/share/ -j
 ...
Scanned 26149 directories and 344654 files,
               processed 72838 inodes,
               328 modified (327 replaced + 1 rewritten),
               0 unsupported format, 1 errors

old: 863.33s user 4.65s system 742% cpu 1:56.94 total
old: 937.24s user 5.06s system 734% cpu 2:08.30 total
new: 835.74s user 4.32s system 738% cpu 1:53.74 total
new: 842.84s user 4.50s system 736% cpu 1:54.98 total
```
So there's some variation, but in general the threaded version seems minimally better.

```console
$ ls -lU target/{debug,release}/add-determinism.{old,new}
 -rwxr-xr-x 1 zbyszek zbyszek 46194504 07-22 20:30 target/debug/add-determinism.old
 -rwxr-xr-x 2 zbyszek zbyszek 45175376 07-22 20:37 target/debug/add-determinism.new
 -rwxr-xr-x 1 zbyszek zbyszek  2434064 07-22 20:30 target/release/add-determinism.old
 -rwxr-xr-x 2 zbyszek zbyszek  2290272 07-22 20:38 target/release/add-determinism.new
```
200kB less in the release version is also nice.